Commit #1- line 73 - i changed it from the AND operation to the OR operation. and i changed the inequality signs and their limits.
Commit #2 - line 79- i changed it from the AND operation to the OR operation. and i changed the inequality signs and their limits.

Commit #3 - lines 85 and 86- i added a new variable called formula to correct the formula used to calculate perimeter

Commit #4 - line 91- i modified the Rvalue of the area variable by separating them into distinct parenthesis